refactor: 源码二次审查优化(复杂度/正确性/一致性)并纳入全部待提交文件
代码优化:
- GcMetrics: GC 类型与 JVM 参数运行期不变,构造时检测一次并缓存,避免诊断周期重复扫描 MXBean
- MSPTCommand: eval(long[]) 的 min 初值 Integer.MAX_VALUE → Long.MAX_VALUE(tick 纳秒值可超 int 范围)
- StartupReporter: replaceAll(正则) → replace(字面量),输出路径无需正则引擎
- ObservabilityBootstrap: 周期任务 warn 日志携带完整异常堆栈(原仅 e.getMessage())
- NetworkMetrics: gauge supplier 统一为方法引用(bytesSent::get 等),与其他收集器一致
- HardwareCapabilities: checkAVX2Support 注释改为准确描述启发式回退行为
- EntityMetrics: Javadoc 措辞修正

审查方式: gen-complexity-scanner skill + Context7 Google Java Style 细则 + 人工热点审查;
风格复扫 35 文件 0 问题;桩编译 exit=0

其他: 纳入 .mcp.json(MCP 配置,无敏感信息)、P1.md、上游参考空文件;移除 AI_HANDOFF.md 与两张过期图片

+++ leviathan-server/src/main/java/dev/vospek/leviathan/command/subcommands/MSPTCommand.java
@@ -215,7 +215,7 @@ public final class MSPTCommand extends PermissionedLeviathanSubcommand {
     }
 
     private static List<Component> eval(long[] times) {
-        long min = Integer.MAX_VALUE;
+        long min = Long.MAX_VALUE;
         long max = 0L;
         long total = 0L;
         int count = 0;
